fix(perfil): incorporar fragmentos de nav-bar
