+++ app/src/main/java/com/kotlin/sacalabici/framework/adapters/views/activities/SettingsActivity.kt
@@ -1,18 +1,17 @@
 package com.kotlin.sacalabici.framework.adapters.views.activities
 
 import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
 import com.kotlin.sacalabici.databinding.ActivitySettingsBinding
 
 
 
-class SettingsActivity: BaseActivity() {
+class SettingsActivity: AppCompatActivity() {
     private lateinit var binding: ActivitySettingsBinding
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-
         initializeBinding()
-        setupNavbar()
     }
 
     private fun initializeBinding(){
